backend-code/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
python-dotenv==1.0.0
Werkzeug==2.3.6
openai==1.3.0
PyPDF2==3.0.1
python-docx==0.8.11

